more sophisticated logic for skipping subspace evaluation
The computation of the inverse covariances is costly,
so if we can skip evaluating Gaussian components which will
not contribute, that saves computation time.

Here the code is working with upper and lower limit estimates
of the gaussian logprob before evaluating the inv cov
to justify skipping components.

This does not work well.

The last part evaluates one component to establish a lower limit
that way.